Fixed echo command bug

+++ commands/echo.command.js
@@ -4,6 +4,11 @@ module.exports = async ctx => {
         const textArray = text.split(" ");
         textArray.splice(0, 1);
         const message = textArray.join(" ");
+        const botData = await ctx.telegram.getChatMember(ctx.chat.id, ctx.botInfo.id);
+
+        if (botData.status !== "administrator" && botData.status !== "creator") {
+            return ctx.reply("To use this feature, the robot must be an administrator and be given access to delete messages.");
+        }
 
         await ctx.deleteMessage();
